remove anonymous and guest separate session logic
